+++ c_api/gpu/GpuIndex_c.h
@@ -22,45 +22,6 @@ FAISS_DECLARE_GETTER(GpuIndexConfig, int, device)
 
 FAISS_DECLARE_CLASS_INHERITED(GpuIndex, Index)
 
-/// Getter for d (dimension)
-FAISS_DECLARE_GETTER(GpuIndex, int, d)
-
-/** Perform training on a representative set of vectors
- *
- * @param index  opaque pointer to GPU index object
- * @param n      nb of training vectors
- * @param x      training vectors, size n * d
- */
-int faiss_GpuIndex_train(FaissGpuIndex* index, idx_t n, const float* x);
-
-/** Add n vectors of dimension d to the index.
- *
- * @param index  opaque pointer to GPU index object
- * @param n      nb of vectors to add
- * @param x      input vectors, size n * d
- */
-int faiss_GpuIndex_add(FaissGpuIndex* index, idx_t n, const float* x);
-
-/** Query n vectors of dimension d to the index.
- *
- * @param index       opaque pointer to GPU index object
- * @param n           number of vectors to search
- * @param x           input vectors to search, size n * d
- * @param k           number of nearest neighbors to return
- * @param distances   output pairwise distances, size n * k
- * @param labels      output labels of the NNs, size n * k
- */
-int faiss_GpuIndex_search(
-        const FaissGpuIndex* index,
-        idx_t n,
-        const float* x,
-        idx_t k,
-        float* distances,
-        idx_t* labels);
-
-/// Free the GPU index
-FAISS_DECLARE_DESTRUCTOR(GpuIndex)
-
 #ifdef __cplusplus
 }
 #endif
